+++ src/features/basic/index.ts
@@ -47,7 +47,7 @@ import './inv-custom-item-sorting/inv-custom-item-sorting';
 import './inv-item-markers/inv-item-markers';
 import './inv-search';
 import './inv-show-space-remaining';
-import './inventory-breakdown';
+import './flt-cargo-breakdown';
 import './item-icons';
 import './item-ticker-shadow';
 import './lead-per-day-column';
